@@ -1,22 +1,24 @@
 {
-  "name": "av-server-ui",
-  "version": "0.4.0",
-  "description": "av-server-ui",
-  "author": "trey",
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "start": "vite --port 4621",
-    "dev": "vite --port 4621 --host",
-    "build": "vite build --emptyOutDir",
-    "preview": "vite preview --port 4621 --host"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.0.1",
-    "svelte": "^3.49.0",
-    "vite": "^3.0.0"
-  },
-  "dependencies": {
-    "svelte-spa-router": "^3.3.0"
-  }
+    "name": "av-server-ui",
+    "version": "0.4.0",
+    "description": "av-server-ui",
+    "author": "trey",
+    "license": "MIT",
+    "type": "module",
+    "scripts": {
+        "start": "vite --port 4621",
+        "dev": "vite --port 4621 --host",
+        "build": "vite build --emptyOutDir",
+        "preview": "vite preview --port 4621 --host"
+    },
+    "devDependencies": {
+        "vite": "^3.2.7",
+        "svelte": "^3.59.2",
+        "@sveltejs/vite-plugin-svelte": "^1.4.0",
+        "vite-plugin-monaco-editor": "^1.1.0"
+    },
+    "dependencies": {
+        "lucide-svelte": "^0.279.0",
+        "monaco-editor": "^0.38.0"
+    }
 }
