@@ -1,241 +1,217 @@
 <!-- Javascript -->
 <script>
-  import { dateObjToDate, dateObjToTime, dateObjToTimeSec } from "../js/helper";
+    import { createEventDispatcher } from "svelte";
+    const dispatch = createEventDispatcher();
 
-  // Components
-  import Icon from './Icon.svelte'
+    // Components
+    import { Circle, X } from 'lucide-svelte';
 
-  // Exports
-  export let showCRLF = true
-  // export let hexSpaces = true
-  export let addTestData = false
-  export let lines = [
-    {
-      wasReceived: false,
-      timestampISO: '2022-10-16T21:05:38.425Z',
-      data: '{"boolean": true, "string": "Yes", "number": 200}',
-    },
-    {
-      wasReceived: true,
-      timestampISO: '2022-10-16T21:05:38.447Z',
-      data: 'OFF',
-    },
-    {
-      wasReceived: false,
-      timestampISO: '2022-10-16T21:05:38.425Z',
-      data: '\x4f\x46\x46',
-    },
-    {
-      wasReceived: true,
-      timestampISO: '2022-10-16T21:05:38.536Z',
-      data: 'Weeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee Weeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee!!!',
-    },
-    {
-      wasReceived: true,
-      timestampISO: '2022-10-16T21:05:38.543Z',
-      data: 'OK',
-    },
-  ]
+    // Exports
+    export let showCRLF = true;
+    export let lines = [
+        {
+            wasReceived: false,
+            timestampISO: "2022-10-16T21:05:38.425Z",
+            data: '{"boolean": true, "string": "Yes", "number": 200}',
+        },
+        {
+            wasReceived: true,
+            timestampISO: "2022-10-16T21:05:38.447Z",
+            data: "OFF",
+        },
+        {
+            wasReceived: false,
+            timestampISO: "2022-10-16T21:05:38.425Z",
+            data: "\x4f\x46\x46",
+        },
+        {
+            wasReceived: true,
+            timestampISO: "2022-10-16T21:05:38.536Z",
+            data: "Weeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee Weeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee!!!",
+        },
+        {
+            wasReceived: true,
+            timestampISO: "2022-10-16T21:05:38.543Z",
+            data: "OK",
+        },
+    ];
+    
+    // Variables
+    let terminalElement;
+    let terminalElementWidth;
+    let linesElement;
 
-  // Variables
-  let terminalElementWidth
-  
-  // Time Column
-  $: timeColFormat = terminalElementWidth > 600 ? "time" : "timeShort"
-  $: timeColWidth = terminalElementWidth > 600 ? "6.25rem" : "5rem"
-  function timePress() {
-    if (timeColFormat === "datetime") {
-      timeColFormat = "time"
-      timeColWidth = "6.25rem"
+    // Time Column
+    $: timeColFormat = terminalElementWidth > 600 ? "time" : "timeShort";
+    $: timeColWidth = terminalElementWidth > 600 ? "8rem" : "4rem";
+    function timePress() {
+        if (timeColFormat === "datetime") {
+            timeColFormat = "time";
+            timeColWidth = "8rem";
+        } else if (timeColFormat === "time") {
+            timeColFormat = "timeShort";
+            timeColWidth = "4rem";
+        } else {
+            timeColFormat = "datetime";
+            timeColWidth = "15rem";
+        }
     }
-    else if (timeColFormat === "time") {
-      timeColFormat = "timeShort"
-      timeColWidth = "5rem"
+
+    // Scroll to the bottom of the Terminal if the scroll bar is at the bottom everytime "lines" updates
+    function scrollToBottomOfElement(element) {
+        const manualOffset = 10
+        const currentScrollHeight = element?.scrollTop + element?.offsetHeight + manualOffset;
+        const totalScrollHeight = element?.scrollHeight;
+        
+        // If the scroll bar is at the bottom
+        if (currentScrollHeight >= totalScrollHeight) {
+            setTimeout(() => (element.scrollTop += 100), 2);
+        }
     }
-    else {
-      timeColFormat = "datetime"
-      timeColWidth = "11.5rem"
+    $: if (lines) scrollToBottomOfElement(linesElement);
+    
+    // Unescape the delimiter if needed
+    $: if (showCRLF) escapeCRLF(lines);
+    function escapeCRLF(lines) {
+        lines.forEach((line) => {
+            line.data = String(line.data).replace(/\r/g, "\\r");
+            line.data = String(line.data).replace(/\n/g, "\\n");
+        });
     }
-  }
-
-  // Max lines in the Terminal
-  let maxLines = 1000
-  $: if (lines.length > maxLines) lines.shift()
-
-  // Scroll to the bottom of the Terminal if the scroll bar is at the bottom everytime "lines" updates
-  let terminalElement
-  function scrollToBottomOfElement(element) {
-    let currentScrollHeight = element?.scrollTop + element?.offsetHeight + 10
-    let totalScrollHeight = element?.scrollHeight
-    // If the scroll bar is at the bottom
-    if (currentScrollHeight >= totalScrollHeight) {
-      element.scrollTop += 100
-      setTimeout(() => element.scrollTop += 100, 5)
-    }
-  }
-  $: if (lines) scrollToBottomOfElement(terminalElement)
-  
-  // Unescape the delimiter if needed
-  $: if (showCRLF) escapeCRLF(lines)
-  function escapeCRLF(lines) {
-    lines.forEach(line => {
-      line.data = String(line.data).replace(/\r/g, "\\r")
-      line.data = String(line.data).replace(/\n/g, "\\n")
-    })
-  }
-
-  // Test Received Data
-  if (addTestData) {
-    let testSends = false
-    setInterval(() => {
-      if (testSends) {
-        lines.push({
-          wasReceived: lines.length%4,
-          timestampISO: new Date(Date.now()).toISOString(),
-          data: lines.length * 89 * 53,
-        })
-        lines = lines
-      }
-    }, 100)
-    setInterval(() => testSends = !testSends , 4 * 1000 )
-  }
-
 </script>
 
 <!-- HTML -->
 <section
-  class="terminal"
-  style="--col2-width: {timeColWidth}"
-  bind:this={terminalElement}
-  bind:offsetWidth={terminalElementWidth}
+    class="terminal"
+    style="--col2-width: {timeColWidth};"
+    bind:this={terminalElement}
+    bind:offsetWidth={terminalElementWidth}
 >
-
-  <!-- Header -->
-  <div class="terminal-header">
-    <div class="terminal-col1"><Icon name="up-down" size=.8 style="display: inline;"/></div>
-    <button class="terminal-col2 textButton" on:click={timePress}>
-      {timeColFormat === "datetime" ? "Date Time" : "Time"}
-      <Icon name="arrows-rotate" size=.8 style="display: inline;" color="var(--color-text-dim)"/>
-    </button>
-    <pre class="terminal-col3">Sent <Icon name="up-long" size=.8 style="display: inline;" color="var(--color-bg-green)"/> / Received <Icon name="down-long" size=.8 style="display: inline;" color="var(--color-bg-blue)"/></pre>
-  </div>
-
-  <!-- Lines -->
-  <div class="terminal-lines">
-    {#each lines as line}
-      <div class="terminal-line textButton">
-
-        <!-- Col1 -->
-        {#if line.wasReceived}
-          <div class="terminal-col1">
-            <Icon name="down-long" size=.8 style="display: inline;" color="var(--color-bg-blue)"/>
-          </div>
-        {:else}
-          <div class="terminal-col1">
-            <Icon name="up-long" size=.8 style="display: inline;" color="var(--color-bg-green)"/>
-          </div>
-        {/if}
-
-        <!-- Col2 -->
-        <div class="terminal-col2">
-          {#if timeColFormat === "datetime"}
-            <span>{dateObjToDate(new Date(line.timestampISO))}</span>
-            <span>{dateObjToTimeSec(new Date(line.timestampISO))}</span>
-          {:else if timeColFormat === "time"}
-            <span>{dateObjToTimeSec(new Date(line.timestampISO))}</span>
-          {:else}
-            <span>{dateObjToTime(new Date(line.timestampISO))}</span>
-          {/if}
+    <!-- Header -->
+    <div class="header">
+        <div class="col1"></div>
+        <button class="col2 textButton" on:click={timePress}>
+            {timeColFormat === "datetime" ? "Date Time" : "Time"}
+        </button>
+        <div class="col3 flex nowrap gap-sm align-center">
+            <span>Sent</span>
+            <X size=.8rem strokeWidth=2.5 color="var(--color-text-orange)" />
+            <span>/</span>
+            <span>Received</span>
+            <Circle size=.8rem strokeWidth=2.5 color="var(--color-text-blue)" />
         </div>
+    </div>
 
-        <!-- Col3 -->
-        <div class="terminal-col3">
-          <pre>{line.data}</pre>
-        </div>
+    <!-- Lines -->
+    <div class="lines" bind:this={linesElement}>
+        {#each lines as line}
+            <!-- svelte-ignore a11y-click-events-have-key-events -->
+            <div class="line" class:mark={line.mark} on:click={() => dispatch("lineClick", line)}>
 
-      </div>
-    {/each}
-    
-  </div>
+                <!-- Col1 -->
+                {#if line.wasReceived}
+                    <div class="col1">
+                        <Circle size=.8rem strokeWidth=2.5 color="var(--color-text-blue)" />
+                    </div>
+                {:else}
+                    <div class="col1">
+                        <X size=.8rem strokeWidth=2.5 color="var(--color-text-orange)" />
+                    </div>
+                {/if}
+
+                <!-- Col2 -->
+                <div class="col2">
+                    {#if timeColFormat === "datetime"}
+                        <span>
+                            {line.timestampISO.split("T")[0]}
+                        </span>
+                        <span>
+                            {line.timestampISO.split("T")[1].split("Z")[0]}
+                        </span>
+                    {:else if timeColFormat === "time"}
+                        <span>
+                            {line.timestampISO.split("T")[1].split("Z")[0]}
+                        </span>
+                    {:else}
+                        <span>
+                            {line.timestampISO.split("T")[1].slice(3).split(".")[0]}
+                        </span>
+                    {/if}
+                </div>
+
+                <!-- Col3 -->
+                <div class="col3">
+                    <pre>{line.data}</pre>
+                </div>
+
+            </div>
+        {/each}
+    </div>
 </section>
 
 <!-- CSS -->
 <style>
-  section {
-    padding: 0;
-    max-height: 100vh;
-  }
-  @media (max-width: 60rem) {
-    section {
-      padding: 0;
-      max-height: 40vh;
+    .terminal {
+        height: 100%;
+        background-color: var(--color-bg);
+        font-family: var(--font-family-mono);
+        font-size: var(--font-family-mono);
+        display: grid;
+        align-content: flex-start;
     }
-  }
-  .terminal {
-    background-color: var(--color-bg);
-    /* border: var(--border); */
-    /* border-color: var(--color-border); */
-    font-family: var(--font-mono);
-    font-size: .8rem;
-    display: grid;
-    align-content: flex-start;
-    overflow: auto;
-    /* resize: both; */
-  }
+    
+    /* Header */
+    .header {
+        padding: var(--gap-sm);
+        display: flex;
+        align-items: center;
 
-  /* Header */
-  .terminal-header {
-    display: flex;
-    align-items: flex-start;
-    color: var(--color-text);
-    padding: var(--pad);
-    border-bottom: var(--border);
-    position: sticky;
-    top: 0;
-    background-color: var(--color-bg);
-    z-index: 1;
-    max-height: 2.5rem;
-  }
+        color: var(--color-text);
+        background-color: var(--color-bg);
+        
+        border-bottom: var(--border);
+    }
+    
+    /* Lines */
+    .lines {
+        padding: var(--gap-sm);
+        min-height: 4rem;
+        height: 100%;
+        overflow: auto;
+    }
+    .line {
+        display: flex;
+        flex-wrap: nowrap;
+        align-items: center;
+    }
+    .line:hover,
+    .line:active,
+    .line:focus {
+        background-color: var(--color-bg-section);
+        filter: var(--filter-brightness-hover);
+    }
 
-  /* Lines */
-  .terminal-lines {
-    padding: var(--pad);
-    padding-top: var(--pad);
-    padding-bottom: var(--pad);
-    min-height: 10rem;
-    /* max-height: 40vh; */
-  }
-  .terminal-line {
-    display: flex;
-    flex-wrap: nowrap;
-    width: 100%;
-  }
-  .terminal-line:hover,
-  .terminal-line:active,
-  .terminal-line:focus {
-    background-color: var(--color-bg-header);
-    filter: var(--filter-brightness-hover);
-  }
+    /* Columns */
+    .col1 {
+        min-width: 1.5rem;
+        display: flex;
+        align-items: center;
+    }
+    .col2 {
+        min-width: var(--col2-width);
+        color: var(--color-text-dim);
+    }
 
-  /* Columns */
-  .terminal-col1 {
-    min-width: 1.5rem;
-  }
-  .terminal-col2 {
-    min-width: var(--col2-width);
-    color: var(--color-text-dim);
-  }
-  .terminal-col2 {
-    min-width: var(--col2-width);
-    color: var(--color-text-dim);
-  }
-
-  .textButton {
-    padding: 0;
-    background-color: transparent;
-    text-align: left;
-    color: currentColor;
-    border-radius: 0;
-  }
-
+    /* Buttons */
+    .textButton {
+        padding: 0;
+        background-color: transparent;
+        text-align: left;
+        color: currentColor;
+        border-radius: 0;
+        border: 0;
+    }
+    .mark {
+        background-color: var(--color-bg-yellow);
+    }
 </style>
