@@ -1,27 +1,36 @@
 import { defineConfig } from 'vite'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
 
-// https://vitejs.dev/config/
+const prefix = 'monaco-editor/esm/vs'
+
 export default defineConfig({
-  base: './',
-  plugins: [svelte()],
-  build: {outDir: '../av-server/public/av-server-ui'}
+base: './',
+    plugins: [svelte()],
+    build: {
+        outDir: '../av-server/server/frontend',
+        rollupOptions: {
+            output: {
+                manualChunks: {
+                    jsonWorker: [`${prefix}/language/json/json.worker`],
+                    cssWorker: [`${prefix}/language/css/css.worker`],
+                    htmlWorker: [`${prefix}/language/html/html.worker`],
+                    tsWorker: [`${prefix}/language/typescript/ts.worker`],
+                    editorWorker: [`${prefix}/editor/editor.worker`],
+                },
+            },
+        },
+    }
 })
 
 /*
-
 Folder Structure expected to build right into av-server
 
-path_to_your_folders
+/path_to_your_folders
 ├── av-server
-│   ├── private
-│   ├── public
-│   │   └── av-server-ui
-│   └── server
+│   ├── server
+│   │   └── frontend
 └── av-server-ui
     ├── public
     ├── src
-    └-- vite.config.js
-
+    └── vite.config.js
 */
-
